Reorganized project in folders to increase clarity; Renamed some classes without the TeamCity prefix, Created TeamCity API wrapper class and started to centralize calls in it, Created entity classes for TC Build and BuildTypes Rewritten logic for QueueBuild and ImportArtifact operations to increase visibility on TC activity, and simplify the usage of the extension Replaced ProjectName/BuildConfig names by Build configuration selection dropdown in plan Operations Commented the post-build VS project statements for those who don't want them, Added two simple cmd files to automatize manual deployment on local dev workstation (without BM),
