+++ Operations/Operation.cs
@@ -1,0 +1,36 @@
+﻿using System.ComponentModel;
+using Inedo.BuildMaster.Extensibility;
+using Inedo.BuildMaster.Extensibility.Credentials;
+using Inedo.BuildMaster.Extensibility.Operations;
+using Inedo.BuildMasterExtensions.TeamCity.Credentials;
+using Inedo.Documentation;
+
+namespace Inedo.BuildMasterExtensions.TeamCity.Operations
+{
+    public abstract class Operation : ExecuteOperation, IHasCredentials<Credentials.Credentials>, IConnectionInfo
+    {
+        [Category("Connection/Identity")]
+        [ScriptAlias("Credentials")]
+        [DisplayName("Credentials")]
+        public string CredentialName { get; set; }
+
+        [Category("Connection/Identity")]
+        [ScriptAlias("Server")]
+        [DisplayName("TeamCity server URL")]
+        [MappedCredential(nameof(Credentials.Credentials.ServerUrl))]
+        public string ServerUrl { get; set; }
+
+        [Category("Connection/Identity")]
+        [ScriptAlias("UserName")]
+        [DisplayName("User name")]
+        [PlaceholderText("Use guest authentication")]
+        [MappedCredential(nameof(Credentials.Credentials.UserName))]
+        public string UserName { get; set; }
+
+        [Category("Connection/Identity")]
+        [ScriptAlias("Password")]
+        [DisplayName("Password")]
+        [MappedCredential(nameof(Credentials.Credentials.Password))]
+        public string Password { get; set; }
+    }
+}
