+++ TeamCity.csproj
@@ -90,55 +90,59 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="BuildNumberPicker.cs" />
-    <Compile Include="Credentials\TeamCityCredentials.cs" />
-    <Compile Include="ITeamCityConnectionInfo.cs" />
-    <Compile Include="Operations\ImportTeamCityArtifactOperation.cs" />
-    <Compile Include="Operations\QueueTeamCityBuildOperation.cs" />
-    <Compile Include="Operations\TeamCityOperation.cs" />
-    <Compile Include="SelectBuildConfigurationPicker.cs" />
-    <Compile Include="GetArtifactAction.cs" />
-    <Compile Include="GetArtifactActionEditor.cs">
+    <Compile Include="Actions\GetArtifact\GetArtifactAction.cs" />
+    <Compile Include="Actions\GetArtifact\GetArtifactActionEditor.cs">
       <SubType>ASPXCodeBehind</SubType>
     </Compile>
-    <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SelectBuildConfigurationVariable.cs" />
-    <Compile Include="SelectBuildConfigurationVariableSetter.cs" />
-    <Compile Include="TeamCityActionBase.cs" />
-    <Compile Include="TeamCityArtifactImporter.cs" />
-    <Compile Include="TeamCityBuildImporter.cs" />
-    <Compile Include="TeamCityBuildImporterEditor.cs">
+    <Compile Include="Actions\DefaultActionBase.cs" />
+    <Compile Include="Actions\TriggerBuild\TriggerBuildAction.cs" />
+    <Compile Include="Actions\TriggerBuild\TriggerBuildActionEditor.cs">
       <SubType>ASPXCodeBehind</SubType>
     </Compile>
-    <Compile Include="TeamCityBuildImporterTemplate.cs" />
-    <Compile Include="TeamCityBuildImporterTemplateEditor.cs">
+    <Compile Include="BuildImporters\BuildImporter.cs" />
+    <Compile Include="BuildImporters\BuildImporterEditor.cs">
       <SubType>ASPXCodeBehind</SubType>
     </Compile>
-    <Compile Include="TeamCityBuildQueuer.cs" />
-    <Compile Include="TeamCityConfigurer.cs" />
-    <Compile Include="TeamCityConfigurerEditor.cs">
+    <Compile Include="BuildImporters\BuildImporterTemplate.cs" />
+    <Compile Include="BuildImporters\BuildImporterTemplateEditor.cs">
       <SubType>ASPXCodeBehind</SubType>
     </Compile>
-    <Compile Include="TeamCityWebClient.cs">
+    <Compile Include="Connectivity\Configurer.cs" />
+    <Compile Include="Connectivity\ConfigurerEditor.cs">
+      <SubType>ASPXCodeBehind</SubType>
+    </Compile>
+    <Compile Include="Connectivity\IConnectionInfo.cs" />
+    <Compile Include="Connectivity\WebClient.cs">
       <SubType>Component</SubType>
     </Compile>
-    <Compile Include="TriggerBuildAction.cs" />
-    <Compile Include="TriggerBuildActionEditor.cs">
-      <SubType>ASPXCodeBehind</SubType>
-    </Compile>
+    <Compile Include="Connectivity\Credentials.cs" />
+    <Compile Include="Controls\BuildConfigurationSelectList.cs" />
+    <Compile Include="Controls\BuildNumberHiddenField.cs" />
+    <Compile Include="Operations\BuildConfigurationArgumentEditor.cs" />
+    <Compile Include="Operations\ImportArtifact\ImportArtifactOperation.cs" />
+    <Compile Include="Operations\ImportArtifact\ImportArtifactOperationEditor.cs" />
+    <Compile Include="Operations\ImportArtifact\ImportArtifactOperationManager.cs" />
+    <Compile Include="Operations\QueueBuild\QueueBuildOperation.cs" />
+    <Compile Include="Operations\QueueBuild\QueueBuildOperationManager.cs" />
+    <Compile Include="Operations\Operation.cs" />
+    <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="TeamCityAPI\TeamCityAPI.cs" />
+    <Compile Include="TeamCityAPI\TeamCityBuild.cs" />
+    <Compile Include="TeamCityAPI\TeamCityBuildType.cs" />
+    <Compile Include="TemplateVariables\TeamCityBuildConfigurationVariable.cs" />
+    <Compile Include="TemplateVariables\TeamCityBuildConfigurationVariableSetter.cs" />
   </ItemGroup>
   <ItemGroup>
-    <Content Include="BuildNumberPicker.js">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
+    <Content Include="Assets\Scripts\BuildNumberPicker.js" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
-    <PostBuildEvent>if exist $(SolutionDir) cmd /c "del $(SolutionDir)..\..\BuildMaster\ExtensionLib\$(TargetName).bmx"
-if exist $(SolutionDir) $(SolutionDir)..\zip -j $(SolutionDir)..\..\BuildMaster\ExtensionLib\$(TargetName).bmx *.*</PostBuildEvent>
+    <PostBuildEvent>REM Uncomment the following if required
+REM if exist $(SolutionDir) cmd /c "del $(SolutionDir)..\..\BuildMaster\ExtensionLib\$(TargetName).bmx"
+REM if exist $(SolutionDir) $(SolutionDir)..\zip -j $(SolutionDir)..\..\BuildMaster\ExtensionLib\$(TargetName).bmx *.*</PostBuildEvent>
   </PropertyGroup>
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
